[refactor] req와 res 폴더 분리
