+++ lib/create/data/mapper/input_content_mapper.dart
@@ -1,5 +1,5 @@
-import 'package:mongo_ai/create/data/dto/input_content_dto.dart';
-import 'package:mongo_ai/create/domain/model/input_content.dart';
+import 'package:mongo_ai/create/data/dto/request/input_content_dto.dart';
+import 'package:mongo_ai/create/domain/model/request/input_content.dart';
 
 extension InputContentMapper on InputContent {
   InputContentDto toInputContentDto() {
