+++ lib/create/data/mapper/message_input_mapper.dart
@@ -1,6 +1,6 @@
-import 'package:mongo_ai/create/data/dto/message_input_dto.dart';
+import 'package:mongo_ai/create/data/dto/request/message_input_dto.dart';
 import 'package:mongo_ai/create/data/mapper/input_content_mapper.dart';
-import 'package:mongo_ai/create/domain/model/message_input.dart';
+import 'package:mongo_ai/create/domain/model/request/message_input.dart';
 
 extension MessageInputMapper on MessageInput {
   MessageInputDto toMessageInputDto() {
