+++ lib/create/data/mapper/open_ai_body_mapper.dart
@@ -1,6 +1,6 @@
-import 'package:mongo_ai/create/data/dto/open_ai_body_dto.dart';
+import 'package:mongo_ai/create/data/dto/request/open_ai_body_dto.dart';
 import 'package:mongo_ai/create/data/mapper/message_input_mapper.dart';
-import 'package:mongo_ai/create/domain/model/open_ai_body.dart';
+import 'package:mongo_ai/create/domain/model/request/open_ai_body.dart';
 
 extension OpenAiBodyMapper on OpenAiBody {
   OpenAIBodyDto toOpenAiBodyDto() {
